requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
openpyxl>=3.1.0
weasyprint>=60.0
bcrypt>=4.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0